pyproject.toml
[tool.poetry]
name = "repl_python3_RightTragicCertification"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^2.0.0"
requests = "^2.25.1"
numpy = "^1.20.3"
Pillow = "^8.2.0"
keras = "^2.4.3"
six = "^1.16.0"
cv2-tools = "^2.4.0"
opencv-contrib-python = "^4.5.2"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
